Add patching for older rails versions
